+++ API/src/main/java/fr/maxlego08/essentials/api/home/HomeUsageType.java
@@ -1,0 +1,8 @@
+package fr.maxlego08.essentials.api.home;
+
+public enum HomeUsageType {
+
+    MAX,
+    STACK
+
+}