+++ src/main/java/fr/maxlego08/essentials/commands/commands/utils/CommandHeal.java
@@ -21,7 +21,7 @@ public class CommandHeal extends VCommand {
 
         Player player = this.argAsPlayer(0, this.player);
 
-        if (this.player == null) return CommandResultType.SYNTAX_ERROR;
+        if (player == null) return CommandResultType.SYNTAX_ERROR;
 
         if (player != this.player && !hasPermission(sender, Permission.ESSENTIALS_HEAL_OTHER)) {
             player = this.player;